pyproject.toml
[tool.poetry]
name = "llm-agent-server"
version = "0.1.0"
description = ""
authors = ["Ahmed Bahloul <abahlou@amazon.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.3"
langchain = "^0.3.4"
google-generativeai = "^0.8.3"
uvicorn = "^0.32.0"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
pylint = "^3.3.1"
mypy = "^1.13.0"
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
